Image submission working

+++ app/Http/Controllers/ProductsController.php
@@ -5,9 +5,19 @@ namespace App\Http\Controllers;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Http\Request;
 use App\Product;
+use Illuminate\Support\Facades\Storage;
 
 class ProductsController extends Controller
 {
+    /**
+     * Create a new controller instance.
+     *
+     * @return void
+     */
+    public function __construct()
+    {
+        $this->middleware('auth', ['except' => ['index', 'show']]);
+    }
     /**
      * Display a listing of the resource.
      *
@@ -81,7 +91,25 @@ class ProductsController extends Controller
             'price' => 'required',
             'state' => 'required',
             'stock' => 'required',
+            'image' => 'image|nullable|max:1999'
         ]);
+
+        // Handle File Upload
+        if($request->hasFile('image')){
+            // Get just filename
+            $filenameWithExtension = $request->file('image')->getClientOriginalName();
+            // Just filename
+            $filename = pathinfo($filenameWithExtension, PATHINFO_FILENAME);
+            // Get just extension
+            $extension = $request->file('image')->getClientOriginalExtension();
+
+            // To store
+            $fileNameToStore = $filename.'_'.time().'.'.$extension;
+            // Upload
+            $path = $request->file('image')->storeAs('public/product_images', $fileNameToStore);
+        }else {
+            $fileNameToStore = 'noimage.jpg';
+        }
         // Create Product
         $product = new Product;
         $product->name = $request->input('name');
@@ -89,6 +117,7 @@ class ProductsController extends Controller
         $product->price = $request->input('price');
         $product->state = $request->input('state');
         $product->stock = $request->input('stock');
+        $product->image_path = $fileNameToStore;
         $product->save();
         return redirect('/products')->with('success', 'Product Created');
     }
@@ -132,7 +161,24 @@ class ProductsController extends Controller
             'price' => 'required',
             'state' => 'required',
             'stock' => 'required',
+            'image' => 'image|nullable|max:1999'
         ]);
+
+        // Handle File Upload
+        if($request->hasFile('image')){
+            // Get just filename
+            $filenameWithExtension = $request->file('image')->getClientOriginalName();
+            // Just filename
+            $filename = pathinfo($filenameWithExtension, PATHINFO_FILENAME);
+            // Get just extension
+            $extension = $request->file('image')->getClientOriginalExtension();
+
+            // To store
+            $fileNameToStore = $filename.'_'.time().'.'.$extension;
+            // Upload
+            $path = $request->file('image')->storeAs('public/product_images', $fileNameToStore);
+        }
+
         // Create Product
         $product = Product::find($id);
         $product->name = $request->input('name');
@@ -140,6 +186,9 @@ class ProductsController extends Controller
         $product->price = $request->input('price');
         $product->state = $request->input('state');
         $product->stock = $request->input('stock');
+        if($request->hasFile('image')){
+            $product->image_path = $fileNameToStore;
+        }
         $product->save();
         return redirect('/products')->with('success', 'Product Updated');
     }
@@ -152,7 +201,14 @@ class ProductsController extends Controller
      */
     public function destroy($id)
     {
+        if(!auth()->user()->isAdmin){
+            return redirect('/products')->with('error', 'Unauthorized Request');
+        }
         $product = Product::find($id);
+        if($product->image_path != 'noimage.jpg'){
+            // Delete Image
+            Storage::delete('public/product_images/'.$product->image_path);
+        }
         $product->delete();
         return redirect('/products')->with('success', 'Deleted Product');
     }
